fixed creating tags folders if uncompress fails.

+++ make-tags.sh
@@ -11,12 +11,14 @@ git tag | while read tag ; do
 	echo "https://github.com/hradec/pipeVFX/archive/$tag.tar.gz"
 	if [ ! -e $tag ] ; then
 		curl -L "https://github.com/hradec/pipeVFX/archive/$tag.tar.gz" | \
-			tar xzv --exclude="golaem" --xform="s/pipeVFX-......pipeline.tools/$tag/x"  pipeVFX-$tag/pipeline/tools
+		tar xzv --exclude="golaem" --xform="s/pipeVFX-......pipeline.tools/$tag/x"  pipeVFX-$tag/pipeline/tools
+		if [ $? -eq 0 ] ; then
 			mv pipeVFX-$tag/pipeline/tools $tag
 			rm -rf pipeVFX-$tag
 
 			# upload to server
 			sudo rsync -avpP ./$tag/ $server:/atomo/pipeline/tags/$tag/
 			ssh 192.168.0.12 " cd /atomo/pipeline/tags/ ; sudo rm -f latest ; sudo ln -s $(ls | egrep '.*\..*\..*' |  sort  -V | tail -1) latest "
+		fi
 	fi
 done
